Add retry and delay properties to support setup cleanup
On a DR cleanup scenario all storage domains are moved to maintenance
and detached from the DC.
Part of those storage domains might run tasks and could not be moved to
maintenance.
In that case the clean process should wait and try to move the SD to
maintenance after a delay.
To overcome that usecase two default attributes were added to configure
the retry and delay properties to support the cleanup.

Bug-Url: https://bugzilla.redhat.com/1533427