Support a consistent file for running vms info on failback
If the file does not exists the process should fetch the running
VMs and save their info in the file.
If the file does exist, the failback process should use the existing file
since we assume that a failback process was already running before and
stopped at the middle.
This patch also add a validation step to validate before failback that
there is no file with running VMs info